#47 [Refactoring] All Repository Queries existing in the Service Layer were moved to the Repository Layer and refactoring was completed.
