+++ src/main/java/com/dessert/gallery/repository/ChatMessage/ChatMessageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long>, ChatMessageRepositoryCustom {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
 
     List<MessageStatusDto> findByChatRoomIdAndDateTimeBetweenOrderByDateTimeDesc(Long chatRoomId, String start, String end);
 }
