--- src/main/java/com/dessert/gallery/repository/ChatMessage/ChatMessageRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.dessert.gallery.repository.ChatMessage;
-
-import com.dessert.gallery.dto.chat.MessageStatusDto;
-
-import java.util.List;
-
-public interface ChatMessageRepositoryCustom {
-
-}
