--- src/main/java/com/dessert/gallery/repository/ChatMessage/ChatMessageRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.dessert.gallery.repository.ChatMessage;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class ChatMessageRepositoryImpl implements ChatMessageRepositoryCustom {
-
-    private final JPAQueryFactory jpaQueryFactory;
-}
