+++ src/main/java/com/dessert/gallery/repository/File/FileRepository.java
@@ -4,7 +4,7 @@ import com.dessert.gallery.entity.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface FileRepository extends JpaRepository<File, Long>, FileRepositoryCustom {
+public interface FileRepository extends JpaRepository<File, Long> {
     List<File> findByNoticeBoard(NoticeBoard notice);
 
     List<File> findByReviewBoard(ReviewBoard review);
