--- src/main/java/com/dessert/gallery/repository/File/FileRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.dessert.gallery.repository.File;
-
-import com.dessert.gallery.entity.*;
-
-import java.util.List;
-
-public interface FileRepositoryCustom {
-
-}
