--- src/main/java/com/dessert/gallery/repository/File/FileRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.dessert.gallery.repository.File;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class FileRepositoryImpl implements FileRepositoryCustom {
-
-    private final JPAQueryFactory jpaQueryFactory;
-}
