+++ src/main/java/com/dessert/gallery/repository/Subscribe/SubscribeRepository.java
@@ -1,12 +1,14 @@
 package com.dessert.gallery.repository.Subscribe;
 
+import com.dessert.gallery.dto.follow.FollowResponseDto;
 import com.dessert.gallery.entity.Store;
 import com.dessert.gallery.entity.Subscribe;
 import com.dessert.gallery.entity.User;
+import com.querydsl.jpa.impl.JPAQuery;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
+import java.util.List;
+
 public interface SubscribeRepository extends JpaRepository<Subscribe, Long>, SubscribeRepositoryCustom {
 
     boolean existsByStoreAndUserAndDeletedIsFalse(Store store, User user);
@@ -18,4 +20,8 @@ public interface SubscribeRepository extends JpaRepository<Subscribe, Long>, Sub
     boolean existsByStoreAndUser(Store store, User user);
 
     Long countAllByStoreAndDeletedIsFalse(Store store);
+
+    List<FollowResponseDto> findDistinctFollowResponseDtoByEmail(int page, String email);
+
+    JPAQuery<Long> countByUserEmailAndDeletedIsFalseAndNothingBlackList(String email);
 }
