+++ src/main/java/com/dessert/gallery/repository/Subscribe/SubscribeRepositoryCustom.java
@@ -1,9 +1,13 @@
 package com.dessert.gallery.repository.Subscribe;
 
-import com.dessert.gallery.entity.Store;
-import com.dessert.gallery.entity.Subscribe;
-import com.dessert.gallery.entity.User;
+import com.dessert.gallery.dto.follow.FollowResponseDto;
+import com.querydsl.jpa.impl.JPAQuery;
+
+import java.util.List;
 
 public interface SubscribeRepositoryCustom {
 
+    List<FollowResponseDto> findDistinctFollowResponseDtoByEmail(int page, String email);
+
+    JPAQuery<Long> countByUserEmailAndDeletedIsFalseAndNothingBlackList(String email);
 }
