--- src/main/java/com/dessert/gallery/repository/User/UserRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.dessert.gallery.repository.User;
-
-public interface UserRepositoryCustom {
-}
