--- src/main/java/com/dessert/gallery/repository/User/UserRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.dessert.gallery.repository.User;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class UserRepositoryImpl implements UserRepositoryCustom {
-
-    private final JPAQueryFactory jpaQueryFactory;
-}
